# doc/api.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:3000
    description: Local server
info:
  title: WASAPhoto API
  version: 1.0.0
...
paths:
...
  /user/{id}:
    parameters:
    - name: id
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234
    post:
      tags: 
        - Upload photo
      operationId: uploadPhoto
      summary: Upload Photo
      description: |
        The user with id {id} is uploading a photo
      requestBody: 
        content: 
          multipart/form-data: 
            schema: {$ref: '#/components/schemas/photo_post'}
      responses:
        "201": 
           description: Photo uploaded successfully.
           content: 
             multipart/form-data: 
               schema: {$ref: '#/components/schemas/photo_post'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []

    put:
      tags: 
        - Update Username
      operationId: setMyUserName
      summary:  Update UserName
      description: | 
         Update the username of a specific user by its ID
      requestBody: 
        required: true
        content:
          application/json: 
           schema: {$ref: '#/components/schemas/user_name'}
      responses:
        "201": 
           description: UserName updated successfully.
           content: 
              application/json: 
                schema: {$ref: '#/components/schemas/user_name'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []
...
  /user/{uid}/likes/{photoid}:
    parameters:
    - name: uid
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234

    - name: photoid
      in: path
      required: true
      description: this is the photo id that the main user wants to like or unlike
      schema:
        type: integer
        description: e.g., /user/1234/photos/1357/like
    
    post:
      tags:
        - Like a photo
      operationId: likePhoto
      summary: Like a photo
      description: |
        The user {uid} wants to like the photo {photoid}
      responses:
        "201":
          description: User {uid} has liked the photo {photoid}
          content:
            application/json: 
              schema:
                description: object for the liked photo
                type: object
                properties:
                  status:
                    description: Status of the opeartion
                    type: string
                    minLength: 3
                    maxLength: 20
                  message:
                    description: Message of the operation
                    type: string
                    minLength: 3
                    maxLength: 20
                  data: {$ref: '#/components/schemas/object_liked'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []

    delete:
      tags: 
        - Unlike a photo
      operationId: unlikePhoto
      summary: Unlike a photo
      description: |
        The user {uid} wants to unlike the photo {photoid}
      responses:
        "204":
          description: User {uid} has unliked the photo {photoid}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}
      
      security:
        - HTTPBearerAuth: []
...
components:
  responses:
...
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  securitySchemes: 
    HTTPBearerAuth:
      type: http
      scheme: bearer
  schemas:
   user_name:
      description: userName object schema
      type: object
      properties:
        name:
          description: User name structure
          type: string
          example: Vincenzo
          pattern: ^.*?$
          minLength: 3
          maxLength: 16
...
   photo_post:
      description: object for photo information 
      type: object
      properties:
        id:
          description: Photo identifier
          type: integer
        photo:
          $ref: '#/components/schemas/object_photo'
        date_upload:
          description: Photo upload date
          type: string
          format: date-time
          minLength: 19
          maxLength: 19
        num_like:
          description: Number of like
          type: integer
          minimum: 0
        num_comment:
          description: Number of comment
          type: integer
          minimum: 0
   object_photo:
    description: object photo
    type: string
    format: binary
    minLength: 0
    maxLength: 3000000
...
   object_liked:
      description: object for liked photo
      type: object
      properties:
        photoId:
          description: Photo identifier
          type: integer
        Num_like:
          description: Number of likes
          type: integer
...